allow floating point player positions

+++ src/Graphics/Gloss/Extra.hs
@@ -28,8 +28,8 @@ insideGrid cellSizeX cellSizeY cells = translate (-centerX) (-centerY)
     totalSizeV = cellSizeV * dimV
     V centerX centerY = totalSizeV / 2
 
-atGridPos :: Float -> Float -> V Int -> Picture -> Picture
-atGridPos dx dy (V x y) = translate (fromIntegral x * dx) (fromIntegral y * dy)
+atGridPos :: Float -> Float -> V Float -> Picture -> Picture
+atGridPos dx dy (V x y) = translate (x * dx) (y * dy)
 
 pictureGrid :: Float -> Float -> [[Picture]] -> Picture
 pictureGrid _ _ [] = blank
